refactor edge endpoints
