+++ app/controllers/concerns/host_verificator.rb
@@ -30,4 +30,33 @@ module HostValidator
       raise ApplicationController::Forbidden
     end
   end
+
+  def verify_edge_host(options)
+    validate_conjur_account(options[:account])
+    msg = ""
+    raise_excep = false
+    if current_user.kind != 'host'
+      raise_excep = true
+      msg = "User kind is: #{current_user.kind}. Should be: 'host'"
+    elsif current_user.role_id.exclude?("host:edge/edge")
+      raise_excep = true
+      msg = "Role is: #{current_user.role_id}. Should include: 'host:edge/edge'"
+    else
+      role = Role[options[:account] + ':group:edge/edge-hosts']
+      unless role&.ancestor_of?(current_user)
+        raise_excep = true
+        msg = "Curren user is: #{current_user}. should be member of #{role}"
+      end
+    end
+
+    if raise_excep
+      logger.error(
+        Errors::Authorization::EndpointNotVisibleToRole.new(
+          msg
+        )
+      )
+      raise ApplicationController::Forbidden
+    end
+  end
 end
+
